Convert OrganizationAdmin to TS
